+++ RELEASENOTES.md
@@ -4,6 +4,7 @@
 * Yap now requires PHP version 8.0 and higher. [#623]
 * Added support for downloading the list of volunteers in JSON format for a given service body. [#612]
 * Show missed calls also as a percentage of total calls received.
+* Metric endpoint now returns volunteer missed and answered metrics, not exposed to UI yet. [#632]
 * Fixed an issue with reports not recursing more than one level. [#620]
 
 ### 4.1.2 (August 17, 2022)